835 AI Agent/requirements.txt
flask
faiss-cpu
langchain
edi-835-parser
ollama
numpy
python-dotenv